services/telegram/requirements.txt
requests
paho-mqtt
sqlalchemy
pytz
psycopg2-binary